refactor(ir): extract edge-condition exhaustiveness to shared ir tier
Move edgesAreExhaustive (+ its ~15 ir-only helpers and knownExhaustiveSets)
out of validator/lint_reachability.go into ir as exported ir.EdgesExhaustive,
and extractEqualityCondition as ir.ExtractEqualityCondition. validator now
calls the shared functions.

Motivation: the simulator/path-enumerator (simulate package) need the same
exhaustiveness test to route the `else` default correctly, but simulate cannot
import validator (validator already imports simulate → cycle). A single
ir-tier source of truth avoids both the cycle and a drift-prone duplicate.
Behavior-preserving: the full validator suite passes unchanged; new ir tests
cover EdgesExhaustive (known set / partition / complementary / negatives).
