+++ validator/lint_reachability.go
@@ -221,20 +221,6 @@ func hasValidStartAndExit(w *ir.Workflow) bool {
 	return w.Start != "" && w.Exit != "" && w.Node(w.Start) != nil && w.Node(w.Exit) != nil
 }
 
-// knownExhaustiveSets lists value sets that are known to be mutually exhaustive
-// for a given variable. If a node's outgoing edges collectively cover all values
-// in any set for a variable, the conditions are exhaustive.
-var knownExhaustiveSets = map[string][][]string{
-	"ctx.outcome": {
-		{"success", "fail"},
-		{"success", "failure"},
-	},
-	"outcome": {
-		{"success", "fail"},
-		{"success", "failure"},
-	},
-}
-
 // findExhaustiveSources returns a set of node IDs whose outgoing conditional
 // edges form an exhaustive condition set (every branch is guaranteed to match).
 func findExhaustiveSources(w *ir.Workflow) map[string]bool {
@@ -242,7 +228,7 @@ func findExhaustiveSources(w *ir.Workflow) map[string]bool {
 	outgoing := buildOutgoingEdgeMap(w)
 
 	for nodeID, edges := range outgoing {
-		if edgesAreExhaustive(edges) {
+		if ir.EdgesExhaustive(edges) {
 			result[nodeID] = true
 		}
 	}
@@ -256,185 +242,4 @@ func buildOutgoingEdgeMap(w *ir.Workflow) map[string][]*ir.Edge {
 		out[e.From] = append(out[e.From], e)
 	}
 	return out
-}
-
-// edgesAreExhaustive returns true if a set of sibling edges forms an
-// exhaustive condition set. Three detection strategies (any match = exhaustive):
-//
-//  1. Known value sets — e.g., outcome = success + outcome = fail
-//  2. Complete partition — all conditional edges test the same variable
-//     with equality and there are 2+ values (author declares "these are
-//     the only cases")
-//  3. Complementary pair — e.g., "contains X" + "not contains X"
-func edgesAreExhaustive(edges []*ir.Edge) bool {
-	byVar := collectConditionValues(edges)
-	if matchesExhaustiveSet(byVar) {
-		return true
-	}
-	if isCompletePartition(edges, byVar) {
-		return true
-	}
-	return hasComplementaryPair(edges)
-}
-
-// isCompletePartition returns true if every conditional edge tests the same
-// variable with equality and there are 2+ distinct values. This means the
-// author has partitioned all routing on a single variable — the conditions
-// cover all intended cases by construction.
-func isCompletePartition(edges []*ir.Edge, byVar map[string]map[string]bool) bool {
-	if len(byVar) != 1 {
-		return false // conditions span multiple variables
-	}
-	conditionalCount := countConditionalEdges(edges)
-	if conditionalCount < 2 {
-		return false // need at least 2 branches to form a partition
-	}
-	equalityCount := countEqualityEdges(edges)
-	return equalityCount == conditionalCount
-}
-
-// countConditionalEdges returns the number of edges with a condition.
-func countConditionalEdges(edges []*ir.Edge) int {
-	n := 0
-	for _, e := range edges {
-		if e.Condition != nil {
-			n++
-		}
-	}
-	return n
-}
-
-// countEqualityEdges returns the number of edges with simple equality conditions.
-func countEqualityEdges(edges []*ir.Edge) int {
-	n := 0
-	for _, e := range edges {
-		if _, ok := extractEqualityCondition(e); ok {
-			n++
-		}
-	}
-	return n
-}
-
-// collectConditionValues groups equality condition values by variable name.
-func collectConditionValues(edges []*ir.Edge) map[string]map[string]bool {
-	byVar := make(map[string]map[string]bool)
-	for _, e := range edges {
-		cmp, ok := extractEqualityCondition(e)
-		if !ok {
-			continue
-		}
-		if byVar[cmp.Variable] == nil {
-			byVar[cmp.Variable] = make(map[string]bool)
-		}
-		byVar[cmp.Variable][cmp.Value] = true
-	}
-	return byVar
-}
-
-// extractEqualityCondition returns the CondCompare if the edge has a simple
-// equality condition (= or ==), and false otherwise.
-func extractEqualityCondition(e *ir.Edge) (ir.CondCompare, bool) {
-	if !hasCondition(e) {
-		return ir.CondCompare{}, false
-	}
-	cmp, ok := e.Condition.Parsed.(ir.CondCompare)
-	if !ok || !isEqualityOp(cmp.Op) {
-		return ir.CondCompare{}, false
-	}
-	return cmp, true
-}
-
-// hasCondition returns true if the edge has a parsed condition.
-func hasCondition(e *ir.Edge) bool {
-	return e.Condition != nil && e.Condition.Parsed != nil
-}
-
-// isEqualityOp returns true for "=" and "==" operators.
-func isEqualityOp(op string) bool {
-	return op == "=" || op == "=="
-}
-
-// matchesExhaustiveSet returns true if any variable's values cover a known exhaustive set.
-func matchesExhaustiveSet(byVar map[string]map[string]bool) bool {
-	for variable, values := range byVar {
-		if variableIsExhaustive(variable, values) {
-			return true
-		}
-	}
-	return false
-}
-
-// variableIsExhaustive returns true if the given values for a variable
-// cover at least one known exhaustive set.
-func variableIsExhaustive(variable string, values map[string]bool) bool {
-	sets, known := knownExhaustiveSets[variable]
-	if !known {
-		return false
-	}
-	for _, set := range sets {
-		if coversAll(values, set) {
-			return true
-		}
-	}
-	return false
-}
-
-// hasComplementaryPair returns true if any two edges form a complementary pair:
-// one edge has condition "var op val" and another has "not var op val".
-// For example: "ctx.tool_stdout contains all-done" + "ctx.tool_stdout not contains all-done".
-func hasComplementaryPair(edges []*ir.Edge) bool {
-	positives, negatives := classifyConditions(edges)
-	for key := range positives {
-		if negatives[key] {
-			return true
-		}
-	}
-	return false
-}
-
-// classifyConditions separates edge conditions into positive comparisons
-// and negated comparisons, keyed by "variable|op|value".
-func classifyConditions(edges []*ir.Edge) (pos, neg map[string]bool) {
-	pos = make(map[string]bool)
-	neg = make(map[string]bool)
-	for _, e := range edges {
-		if !hasCondition(e) {
-			continue
-		}
-		if key, ok := conditionKey(e.Condition.Parsed); ok {
-			pos[key] = true
-		}
-		if key, ok := negatedConditionKey(e.Condition.Parsed); ok {
-			neg[key] = true
-		}
-	}
-	return pos, neg
-}
-
-// conditionKey returns a string key for a simple comparison condition.
-func conditionKey(expr ir.ConditionExpr) (string, bool) {
-	cmp, ok := expr.(ir.CondCompare)
-	if !ok {
-		return "", false
-	}
-	return cmp.Variable + "|" + cmp.Op + "|" + cmp.Value, true
-}
-
-// negatedConditionKey returns the key for the inner comparison of a CondNot.
-func negatedConditionKey(expr ir.ConditionExpr) (string, bool) {
-	neg, ok := expr.(ir.CondNot)
-	if !ok {
-		return "", false
-	}
-	return conditionKey(neg.Inner)
-}
-
-// coversAll returns true if values contains every element in required.
-func coversAll(values map[string]bool, required []string) bool {
-	for _, r := range required {
-		if !values[r] {
-			return false
-		}
-	}
-	return true
 }
